Add Storage Class
Storage is a class which allows creation and usage of SQLite backed
datastorage. The datastore is accessible via a `WebStorage`-like API,
with a few utility properties (`.keys`, `.values`, `.entries`) added, as
well as `.prepare(query)`, `.transaction(transactionFn)` and
`.exec(query)` to make use of the powerful capabilities of an SQL
backend.